Fix UpdateJobTime.FrameCount to be of type int

+++ Runtime/Jobs/UpdateJobTime.cs
@@ -17,11 +17,12 @@ namespace Gilzoide.UpdateManager.Jobs
         /// <summary>Cached value for <see cref="UnityEngine.Time.unscaledDeltaTime"/> from current running frame</summary>
         public static float unscaledDeltaTime => InstanceRef.UnscaledDeltaTime;
         /// <summary>Cached value for <see cref="UnityEngine.Time.realtimeSinceStartup"/> from current running frame</summary>
+        /// <remarks>Contrary to <see cref="UnityEngine.Time.realtimeSinceStartup"/>, this property will return the same value if called twice during the same frame.</remarks>
         public static float realtimeSinceStartup => InstanceRef.RealtimeSinceStartup;
         /// <summary>Cached value for <see cref="UnityEngine.Time.timeSinceLevelLoad"/> from current running frame</summary>
         public static float timeSinceLevelLoad => InstanceRef.TimeSinceLevelLoad;
         /// <summary>Cached value for <see cref="UnityEngine.Time.frameCount"/> from current running frame</summary>
-        public static float frameCount => InstanceRef.FrameCount;
+        public static int frameCount => InstanceRef.FrameCount;
 
         /// <summary>Cached value for <see cref="UnityEngine.Time.time"/> from current running frame</summary>
         public float Time { get; private set; }
@@ -32,11 +33,12 @@ namespace Gilzoide.UpdateManager.Jobs
         /// <summary>Cached value for <see cref="UnityEngine.Time.unscaledDeltaTime"/> from current running frame</summary>
         public float UnscaledDeltaTime { get; private set; }
         /// <summary>Cached value for <see cref="UnityEngine.Time.realtimeSinceStartup"/> from current running frame</summary>
+        /// <remarks>Contrary to <see cref="UnityEngine.Time.realtimeSinceStartup"/>, this property will return the same value if called twice during the same frame.</remarks>
         public float RealtimeSinceStartup { get; private set; }
         /// <summary>Cached value for <see cref="UnityEngine.Time.timeSinceLevelLoad"/> from current running frame</summary>
         public float TimeSinceLevelLoad { get; private set; }
         /// <summary>Cached value for <see cref="UnityEngine.Time.frameCount"/> from current running frame</summary>
-        public float FrameCount { get; private set; }
+        public int FrameCount { get; private set; }
 
         public static UpdateJobTime Instance => InstanceRef;
